fix: fix bypassing issue
