+++ src/uv_dynamic_versioning/main.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import os
 from pathlib import Path
 
@@ -22,10 +24,11 @@ def validate(project: tomlkit.TOMLDocument):
     return schemas.Project.model_validate(project.unwrap())
 
 
-def _get_version(config: schemas.UvDynamicVersioning) -> Version:
-    if "UV_DYNAMIC_VERSIONING_BYPASS" in os.environ:
-        return Version.parse(os.environ["UV_DYNAMIC_VERSIONING_BYPASS"])
+def _get_bypassed_version() -> str | None:
+    return os.environ.get("UV_DYNAMIC_VERSIONING_BYPASS")
 
+
+def _get_version(config: schemas.UvDynamicVersioning) -> Version:
     try:
         return Version.from_vcs(
             config.vcs,
@@ -46,6 +49,10 @@ def _get_version(config: schemas.UvDynamicVersioning) -> Version:
 
 
 def get_version(config: schemas.UvDynamicVersioning) -> tuple[str, Version]:
+    bypassed = _get_bypassed_version()
+    if bypassed:
+        return bypassed, Version.parse(bypassed)
+
     version = _get_version(config)
 
     if config.format_jinja:
